class_10/pyproject.toml
[project]
name = "class-10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.19",
]
